+++ Bootstrap.php
@@ -99,14 +99,7 @@ class Bootstrap implements BootstrapInterface
                 ];
             }
 
-            $defaults = [
-                'welcomeSubject'        => Yii::t('user', 'Welcome to {0}', Yii::$app->name),
-                'confirmationSubject'   => Yii::t('user', 'Confirm account on {0}', Yii::$app->name),
-                'reconfirmationSubject' => Yii::t('user', 'Confirm email change on {0}', Yii::$app->name),
-                'recoverySubject'       => Yii::t('user', 'Complete password reset on {0}', Yii::$app->name),
-            ];
-
-            Yii::$container->set('dektrium\user\Mailer', array_merge($defaults, $module->mailer));
+            Yii::$container->set('dektrium\user\Mailer', $module->mailer);
         }
     }
 }